+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/tabs/behavior-tab/utils/behavior-to-german-name.pipe.ts
@@ -14,6 +14,8 @@ const behaviorTypeToGermanNameDictionary: {
     automaticallyDistributeVehiclesBehavior: 'Fahrzeuge verteilen',
     requestBehavior: 'Fahrzeuge anfordern',
     transferBehavior: 'Fahrzeuge versenden',
+    transferToHospitalBehavior: 'Patienten abtransportieren',
+    managePatientTransportToHospitalBehavior: 'Transportorganisation',
 };
 @Pipe({
     name: 'behaviorTypeToGermanName',