+++ shared/src/simulation/behaviors/exercise-simulation-behavior.ts
@@ -9,6 +9,8 @@ import { providePersonnelBehavior } from './provide-personnel';
 import { answerRequestsBehavior } from './answer-requests';
 import { requestBehavior } from './request';
 import { transferBehavior } from './transfer';
+import { transferToHospitalBehavior } from './transfer-to-hospital';
+import { managePatientTransportToHospitalBehavior } from './manage-patient-transport-to-hospital';
 
 export const simulationBehaviors = {
     automaticallyDistributeVehiclesBehavior,
@@ -20,6 +22,8 @@ export const simulationBehaviors = {
     answerRequestsBehavior,
     requestBehavior,
     transferBehavior,
+    transferToHospitalBehavior,
+    managePatientTransportToHospitalBehavior,
 };
 
 export type ExerciseSimulationBehavior =