+++ shared/src/simulation/behaviors/request.spec.ts
@@ -366,9 +366,13 @@ const sendEvent = {
         simulatedRegion: Mutable<SimulatedRegion>,
         behaviorState: Mutable<RequestBehaviorState>
     ) => {
+        const transferPoint = Object.values(draftState.transferPoints)[0]!;
         sendSimulationEvent(
             simulatedRegion,
-            VehiclesSentEvent.create(VehicleResource.create({ KTW: 1 }))
+            VehiclesSentEvent.create(
+                VehicleResource.create({ KTW: 1 }),
+                transferPoint.id
+            )
         );
     },
     ktwVehicleArrivedEvent: (