+++ shared/src/simulation/behaviors/transfer.ts
@@ -30,7 +30,10 @@ import {
     RecurringEventActivityState,
     SendRemoteEventActivityState,
 } from '../activities';
-import { isUnoccupied } from '../../models/utils/occupations/occupation-helpers-mutable';
+import {
+    isUnoccupied,
+    isUnoccupiedOrIntermediarilyOccupied,
+} from '../../models/utils/occupations/occupation-helpers-mutable';
 import { amountOfResourcesInVehicle } from '../../models/utils/amount-of-resources-in-vehicle';
 import type { ResourceDescription } from '../../models/utils/resource-description';
 import {
@@ -159,7 +162,12 @@ export const transferBehavior: SimulationBehavior<TransferBehaviorState> = {
                         event.vehicleId
                     );
                     // Don't do anything if vehicle is occupied
-                    if (!isUnoccupied(vehicle, draftState.currentTime)) {
+                    if (
+                        !isUnoccupiedOrIntermediarilyOccupied(
+                            vehicle,
+                            draftState.currentTime
+                        )
+                    ) {
                         return;
                     }
 
@@ -313,22 +321,33 @@ export const transferBehavior: SimulationBehavior<TransferBehaviorState> = {
                         )
                     );
 
-                    // Send event to destination if it is a simulated region
+                    // Send event to transfer initiating region
 
-                    if (
-                        event.transferDestinationType === 'transferPoint' &&
-                        isInSimulatedRegion(
-                            getElement(
-                                draftState,
-                                'transferPoint',
-                                event.transferDestinationId
-                            )
-                        )
-                    ) {
+                    if (event.transferInitiatingRegionId) {
                         addActivity(
                             simulatedRegion,
                             SendRemoteEventActivityState.create(
                                 nextUUID(draftState),
+                                event.transferInitiatingRegionId,
+                                VehiclesSentEvent.create(
+                                    VehicleResource.create(sentVehicles),
+                                    event.transferDestinationId,
+                                    event.key
+                                )
+                            )
+                        );
+                    }
+
+                    // Send event to destination if it is a simulated region and not the initiating region
+                    if (event.transferDestinationType === 'transferPoint') {
+                        const transferPoint = getElement(
+                            draftState,
+                            'transferPoint',
+                            event.transferDestinationId
+                        );
+
+                        if (isInSimulatedRegion(transferPoint)) {
+                            const targetSimulatedRegion =
                                 currentSimulatedRegionOf(
                                     draftState,
                                     getElement(
@@ -336,12 +355,27 @@ export const transferBehavior: SimulationBehavior<TransferBehaviorState> = {
                                         'transferPoint',
                                         event.transferDestinationId
                                     )
-                                ).id,
-                                VehiclesSentEvent.create(
-                                    VehicleResource.create(sentVehicles)
-                                )
-                            )
-                        );
+                                );
+
+                            if (
+                                targetSimulatedRegion.id !==
+                                event.transferInitiatingRegionId
+                            ) {
+                                addActivity(
+                                    simulatedRegion,
+                                    SendRemoteEventActivityState.create(
+                                        nextUUID(draftState),
+                                        targetSimulatedRegion.id,
+                                        VehiclesSentEvent.create(
+                                            VehicleResource.create(
+                                                sentVehicles
+                                            ),
+                                            transferPoint.id
+                                        )
+                                    )
+                                );
+                            }
+                        }
                     }
                 }
                 break;